Move MaybeUtf8 and Parameters into context module
