+++ src/lib.rs
@@ -118,11 +118,6 @@ pub mod file;
 use std::net::{SocketAddr, SocketAddrV4, SocketAddrV6, Ipv4Addr};
 use std::str::FromStr;
 use std::any::TypeId;
-use std::collections::hash_map::{HashMap, Entry};
-use std::ops::{Deref, DerefMut};
-use std::borrow::{Cow, Borrow};
-use std::hash::{Hash, Hasher};
-use std::fmt;
 
 use anymap::Map;
 use anymap::any::{Any, UncheckedAnyExt};
@@ -361,358 +356,4 @@ enum GlobalState {
     None,
     One(TypeId, Box<Any + Send + Sync>),
     Many(Map<Any + Send + Sync>),
-}
-
-///An extended `HashMap` with extra functionality for value parsing.
-///
-///Some of the methods from `HashMap` has been wrapped to provide a more
-///ergonomic API, where anything that can be represented as a byte slice can
-///be used as a key.
-#[derive(Clone)]
-pub struct Parameters(HashMap<MaybeUtf8Owned, MaybeUtf8Owned>);
-
-impl Parameters {
-    ///Create an empty `Parameters`.
-    pub fn new() -> Parameters {
-        Parameters(HashMap::new())
-    }
-
-    ///Get a parameter as a UTF-8 string. A lossy conversion will be performed
-    ///if it's not encoded as UTF-8. Use `get_raw` to get the original data.
-    pub fn get<'a, K: ?Sized>(&'a self, key: &K) -> Option<Cow<'a, str>> where
-        K: Hash + Eq + AsRef<[u8]>
-    {
-        self.0.get(key.as_ref()).map(|v| v.as_utf8_lossy())
-    }
-
-    ///Get a parameter that may or may not be a UTF-8 string.
-    pub fn get_raw<'a, K: ?Sized>(&'a self, key: &K) -> Option<&'a MaybeUtf8Owned> where
-        K: Hash + Eq + AsRef<[u8]>
-    {
-        self.0.get(key.as_ref())
-    }
-
-    ///Get a mutable parameter that may or may not be a UTF-8 string.
-    pub fn get_mut<'a, K: ?Sized>(&'a mut self, key: &K) -> Option<&'a mut MaybeUtf8Owned> where
-        K: Hash + Eq + AsRef<[u8]>
-    {
-        self.0.get_mut(key.as_ref())
-    }
-
-    ///Returns true if a parameter with the given key exists.
-    pub fn contains_key<K: ?Sized>(&self, key: &K) -> bool where
-        K: Hash + Eq + AsRef<[u8]>
-    {
-        self.0.contains_key(key.as_ref())
-    }
-
-    ///Insert a parameter.
-    pub fn insert<K, V>(&mut self, key: K, value: V) -> Option<MaybeUtf8Owned> where
-        K: Into<MaybeUtf8Owned>, V: Into<MaybeUtf8Owned>
-    {
-        self.0.insert(key.into(), value.into())
-    }
-
-    ///Remove a parameter and return it.
-    pub fn remove<K: ?Sized>(&mut self, key: &K) -> Option<MaybeUtf8Owned> where
-        K: Hash + Eq + AsRef<[u8]>
-    {
-        self.0.remove(key.as_ref())
-    }
-
-    ///Gets the given key's corresponding parameter in the map for in-place
-    ///manipulation.
-    pub fn entry<K>(&mut self, key: K) -> Entry<MaybeUtf8Owned, MaybeUtf8Owned> where K: Into<MaybeUtf8Owned> {
-        self.0.entry(key.into())
-    }
-
-    ///Try to parse an entry as `T`, if it exists. The error will be `None` if
-    ///the entry does not exist, and `Some` if it does exists, but the parsing
-    ///failed.
-    ///
-    ///```
-    ///# use rustful::{Context, Response};
-    ///fn my_handler(context: Context, response: Response) {
-    ///    let age: Result<u8, _> = context.variables.parse("age");
-    ///    match age {
-    ///        Ok(age) => response.send(format!("age: {}", age)),
-    ///        Err(Some(_)) => response.send("age must be a positive number"),
-    ///        Err(None) => response.send("no age provided")
-    ///    }
-    ///}
-    ///```
-    pub fn parse<K: ?Sized, T>(&self, key: &K) -> Result<T, Option<T::Err>> where
-        K: Hash + Eq + AsRef<[u8]>,
-        T: FromStr
-    {
-        if let Some(val) = self.0.get(key.as_ref()) {
-            val.as_utf8_lossy().parse().map_err(|e| Some(e))
-        } else {
-            Err(None)
-        }
-    }
-
-    ///Try to parse an entry as `T`, if it exists, or return the default in
-    ///`or`.
-    ///
-    ///```
-    ///# use rustful::{Context, Response};
-    ///fn my_handler(context: Context, response: Response) {
-    ///    let page = context.variables.parse_or("page", 0u8);
-    ///    response.send(format!("current page: {}", page));
-    ///}
-    ///```
-    pub fn parse_or<K: ?Sized, T>(&self, key: &K, or: T) -> T where
-        K: Hash + Eq + AsRef<[u8]>,
-        T: FromStr
-    {
-        self.parse(key).unwrap_or(or)
-    }
-
-    ///Try to parse an entry as `T`, if it exists, or create a new one using
-    ///`or_else`. The `or_else` function will receive the parsing error if the
-    ///value existed, but was impossible to parse.
-    ///
-    ///```
-    ///# use rustful::{Context, Response};
-    ///# fn do_heavy_stuff() -> u8 {0}
-    ///fn my_handler(context: Context, response: Response) {
-    ///    let science = context.variables.parse_or_else("science", |_| do_heavy_stuff());
-    ///    response.send(format!("science value: {}", science));
-    ///}
-    ///```
-    pub fn parse_or_else<K: ?Sized, T, F>(&self, key: &K, or_else: F) -> T where
-        K: Hash + Eq + AsRef<[u8]>,
-        T: FromStr,
-        F: FnOnce(Option<T::Err>) -> T
-    {
-        self.parse(key).unwrap_or_else(or_else)
-    }
-}
-
-impl Deref for Parameters {
-    type Target = HashMap<MaybeUtf8Owned, MaybeUtf8Owned>;
-
-    fn deref(&self) -> &HashMap<MaybeUtf8Owned, MaybeUtf8Owned> {
-        &self.0
-    }
-}
-
-impl DerefMut for Parameters {
-    fn deref_mut(&mut self) -> &mut HashMap<MaybeUtf8Owned, MaybeUtf8Owned> {
-        &mut self.0
-    }
-}
-
-impl AsRef<HashMap<MaybeUtf8Owned, MaybeUtf8Owned>> for Parameters {
-    fn as_ref(&self) -> &HashMap<MaybeUtf8Owned, MaybeUtf8Owned> {
-        &self.0
-    }
-}
-
-impl AsMut<HashMap<MaybeUtf8Owned, MaybeUtf8Owned>> for Parameters {
-    fn as_mut(&mut self) -> &mut HashMap<MaybeUtf8Owned, MaybeUtf8Owned> {
-        &mut self.0
-    }
-}
-
-impl Into<HashMap<MaybeUtf8Owned, MaybeUtf8Owned>> for Parameters {
-    fn into(self) -> HashMap<MaybeUtf8Owned, MaybeUtf8Owned> {
-        self.0
-    }
-}
-
-impl From<HashMap<MaybeUtf8Owned, MaybeUtf8Owned>> for Parameters {
-    fn from(map: HashMap<MaybeUtf8Owned, MaybeUtf8Owned>) -> Parameters {
-        Parameters(map)
-    }
-}
-
-impl PartialEq for Parameters {
-    fn eq(&self, other: &Parameters) -> bool {
-        self.0.eq(&other.0)
-    }
-}
-
-impl Eq for Parameters {}
-
-impl fmt::Debug for Parameters {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl Default for Parameters {
-    fn default() -> Parameters {
-        Parameters::new()
-    }
-}
-
-impl IntoIterator for Parameters {
-    type IntoIter = <HashMap<MaybeUtf8Owned, MaybeUtf8Owned> as IntoIterator>::IntoIter;
-    type Item = (MaybeUtf8Owned, MaybeUtf8Owned);
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.0.into_iter()
-    }
-}
-
-impl<'a> IntoIterator for &'a Parameters {
-    type IntoIter = <&'a HashMap<MaybeUtf8Owned, MaybeUtf8Owned> as IntoIterator>::IntoIter;
-    type Item = (&'a MaybeUtf8Owned, &'a MaybeUtf8Owned);
-
-    fn into_iter(self) -> Self::IntoIter {
-        (&self.0).into_iter()
-    }
-}
-
-impl<'a> IntoIterator for &'a mut Parameters {
-    type IntoIter = <&'a mut HashMap<MaybeUtf8Owned, MaybeUtf8Owned> as IntoIterator>::IntoIter;
-    type Item = (&'a MaybeUtf8Owned, &'a mut MaybeUtf8Owned);
-
-    fn into_iter(self) -> Self::IntoIter {
-        (&mut self.0).into_iter()
-    }
-}
-
-impl<K: Into<MaybeUtf8Owned>, V: Into<MaybeUtf8Owned>> std::iter::FromIterator<(K, V)> for Parameters {
-    fn from_iter<T: IntoIterator<Item=(K, V)>>(iterable: T) -> Parameters {
-        HashMap::from_iter(iterable.into_iter().map(|(k, v)| (k.into(), v.into()))).into()
-    }
-}
-
-impl<K: Into<MaybeUtf8Owned>, V: Into<MaybeUtf8Owned>> Extend<(K, V)> for Parameters {
-    fn extend<T: IntoIterator<Item=(K, V)>>(&mut self, iter: T) {
-        self.0.extend(iter.into_iter().map(|(k, v)| (k.into(), v.into())))
-    }
-}
-
-///An owned string that may be UTF-8 encoded.
-pub type MaybeUtf8Owned = MaybeUtf8<String, Vec<u8>>;
-///A slice of a string that may be UTF-8 encoded.
-pub type MaybeUtf8Slice<'a> = MaybeUtf8<&'a str, &'a [u8]>;
-
-///String data that may or may not be UTF-8 encoded.
-#[derive(Debug, Clone)]
-pub enum MaybeUtf8<S, V> {
-    ///A UTF-8 encoded string.
-    Utf8(S),
-    ///A non-UTF-8 string.
-    NotUtf8(V)
-}
-
-impl<S, V> MaybeUtf8<S, V> {
-    ///Produce a slice of this string.
-    pub fn as_slice<Sref: ?Sized, Vref: ?Sized>(&self) -> MaybeUtf8<&Sref, &Vref> where S: AsRef<Sref>, V: AsRef<Vref> {
-        match *self {
-            MaybeUtf8::Utf8(ref s) => MaybeUtf8::Utf8(s.as_ref()),
-            MaybeUtf8::NotUtf8(ref v) => MaybeUtf8::NotUtf8(v.as_ref())
-        }
-    }
-
-    ///Borrow the string if it's encoded as valid UTF-8.
-    pub fn as_utf8<'a>(&'a self) -> Option<&'a str> where S: AsRef<str> {
-        match *self {
-            MaybeUtf8::Utf8(ref s) => Some(s.as_ref()),
-            MaybeUtf8::NotUtf8(_) => None
-        }
-    }
-
-    ///Borrow the string if it's encoded as valid UTF-8, or make a lossy conversion.
-    pub fn as_utf8_lossy<'a>(&'a self) -> Cow<'a, str> where S: AsRef<str>, V: AsRef<[u8]> {
-        match *self {
-            MaybeUtf8::Utf8(ref s) => s.as_ref().into(),
-            MaybeUtf8::NotUtf8(ref v) => String::from_utf8_lossy(v.as_ref())
-        }
-    }
-
-    ///Borrow the string as a slice of bytes.
-    pub fn as_bytes(&self) -> &[u8] where S: AsRef<[u8]>, V: AsRef<[u8]> {
-        match *self {
-            MaybeUtf8::Utf8(ref s) => s.as_ref(),
-            MaybeUtf8::NotUtf8(ref v) => v.as_ref()
-        }
-    }
-}
-
-impl<V> From<String> for MaybeUtf8<String, V> {
-    fn from(string: String) -> MaybeUtf8<String, V> {
-        MaybeUtf8::Utf8(string)
-    }
-}
-
-impl<'a, V> From<&'a str> for MaybeUtf8<&'a str, V> {
-    fn from(string: &'a str) -> MaybeUtf8<&'a str, V> {
-        MaybeUtf8::Utf8(string)
-    }
-}
-
-impl From<Vec<u8>> for MaybeUtf8<String, Vec<u8>> {
-    fn from(bytes: Vec<u8>) -> MaybeUtf8<String, Vec<u8>> {
-        match String::from_utf8(bytes) {
-            Ok(string) => MaybeUtf8::Utf8(string),
-            Err(e) => MaybeUtf8::NotUtf8(e.into_bytes())
-        }
-    }
-}
-
-impl<S: AsRef<[u8]>, V: AsRef<[u8]>> AsRef<[u8]> for MaybeUtf8<S, V> {
-    fn as_ref(&self) -> &[u8] {
-        match *self {
-            MaybeUtf8::Utf8(ref s) => s.as_ref(),
-            MaybeUtf8::NotUtf8(ref v) => v.as_ref()
-        }
-    }
-}
-
-impl<S: AsRef<[u8]>, V: AsRef<[u8]>> Borrow<[u8]> for MaybeUtf8<S, V> {
-    fn borrow(&self) -> &[u8] {
-        self.as_ref()
-    }
-}
-
-impl<S: AsRef<[u8]>, V: AsRef<[u8]>> PartialEq for MaybeUtf8<S, V> {
-    fn eq(&self, other: &MaybeUtf8<S, V>) -> bool {
-        self.as_ref().eq(other.as_ref())
-    }
-}
-
-impl<S: AsRef<[u8]>, V: AsRef<[u8]>> Eq for MaybeUtf8<S, V> {}
-
-impl<S: AsRef<[u8]>, V: AsRef<[u8]>> Hash for MaybeUtf8<S, V> {
-    fn hash<H: Hasher>(&self, hasher: &mut H) {
-        self.as_ref().hash(hasher)
-    }
-}
-
-impl<S: Into<String>, V: Into<Vec<u8>>> Into<String> for MaybeUtf8<S, V> {
-    fn into(self) -> String {
-        match self {
-            MaybeUtf8::Utf8(s) => s.into(),
-            MaybeUtf8::NotUtf8(v) => {
-                let bytes = v.into();
-                match String::from_utf8_lossy(&bytes) {
-                    Cow::Borrowed(_) => unsafe { String::from_utf8_unchecked(bytes) },
-                    Cow::Owned(s) => s
-                }
-            }
-        }
-    }
-}
-
-impl<S: Into<Vec<u8>>, V: Into<Vec<u8>>> Into<Vec<u8>> for MaybeUtf8<S, V> {
-    fn into(self) -> Vec<u8> {
-        match self {
-            MaybeUtf8::Utf8(s) => s.into(),
-            MaybeUtf8::NotUtf8(v) => v.into()
-        }
-    }
-}
-
-impl<S: AsRef<[u8]>, V: AsRef<[u8]>> Deref for MaybeUtf8<S, V> {
-    type Target=[u8];
-
-    fn deref(&self) -> &[u8] {
-        self.as_ref()
-    }
 }
