+++ src/router/mod.rs
@@ -87,8 +87,8 @@ use std::ops::Deref;
 use hyper::method::Method;
 
 use handler::Handler;
+use context::MaybeUtf8Owned;
 use context::hypermedia::Hypermedia;
-use MaybeUtf8Owned;
 
 pub use self::tree_router::TreeRouter;
